% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{naive_bayes}
\alias{naive_bayes}
\title{naive_bayes}
\description{
This class implements a supervised learning methods based on applying Bayes’
theorem with strong (naive) feature independence assumption
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-naive_bayes-fit}{\code{naive_bayes$fit()}}
\item \href{#method-naive_bayes-prob}{\code{naive_bayes$prob()}}
\item \href{#method-naive_bayes-predict_}{\code{naive_bayes$predict_()}}
\item \href{#method-naive_bayes-predict}{\code{naive_bayes$predict()}}
\item \href{#method-naive_bayes-predict_proba}{\code{naive_bayes$predict_proba()}}
\item \href{#method-naive_bayes-print}{\code{naive_bayes$print()}}
\item \href{#method-naive_bayes-clone}{\code{naive_bayes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-fit"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$fit(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-prob"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-prob}{}}}
\subsection{Method \code{prob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$prob(label, x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-predict_"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-predict_}{}}}
\subsection{Method \code{predict_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$predict_(x, with_prob = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-predict"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$predict(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$predict_proba(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-print"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-naive_bayes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-naive_bayes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
