% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayesClassifier.R
\name{NaiveBayesClassifier}
\alias{NaiveBayesClassifier}
\title{NaiveBayesClassifier}
\description{
This class implements a supervised learning methods based on applying Bayes’ theorem with strong (naive) feature independence assumption
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NaiveBayesClassifier-fit}{\code{NaiveBayesClassifier$fit()}}
\item \href{#method-NaiveBayesClassifier-prob}{\code{NaiveBayesClassifier$prob()}}
\item \href{#method-NaiveBayesClassifier-predict_}{\code{NaiveBayesClassifier$predict_()}}
\item \href{#method-NaiveBayesClassifier-predict}{\code{NaiveBayesClassifier$predict()}}
\item \href{#method-NaiveBayesClassifier-predict_proba}{\code{NaiveBayesClassifier$predict_proba()}}
\item \href{#method-NaiveBayesClassifier-print}{\code{NaiveBayesClassifier$print()}}
\item \href{#method-NaiveBayesClassifier-clone}{\code{NaiveBayesClassifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$fit(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{features}

\item{\code{y}}{target}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-prob"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-prob}{}}}
\subsection{Method \code{prob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$prob(label, x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict_"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict_}{}}}
\subsection{Method \code{predict_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$predict_(x, with_prob = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$predict(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$predict_proba(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-print"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayesClassifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
