% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_indices_stratified.R
\name{split_indices_stratified}
\alias{split_indices_stratified}
\title{Title}
\usage{
split_indices_stratified(labels, train_size)
}
\arguments{
\item{labels}{an array of labels to use to stratify data}

\item{train_size}{a number between 0 and 1 portion of number to include in train set}
}
\value{
indexes of train samples
}
\description{
Title
}
