% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{train_test_split}
\usage{
train_test_split(data, train_size = 0.7, stratify = NULL, seed = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{train_size}{a number between 0 and 1 portion of number to include in train set}

\item{stratify}{a character the name of column to be use to strafied data}

\item{seed}{a number to fixed sample generator}
}
\description{
train_test_split
}
\examples{
train_test_split(iris, train_size = 0.7, stratify = "Species", seed = 123)
}
