% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Classifier}
\alias{Classifier}
\title{Une classe R6 pour un classificateur simple.}
\description{
Cette classe simule un classificateur avec des méthodes d'entraînement et de prédiction.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Classifier-new}{\code{Classifier$new()}}
\item \href{#method-Classifier-entrainer}{\code{Classifier$entrainer()}}
\item \href{#method-Classifier-predire}{\code{Classifier$predire()}}
\item \href{#method-Classifier-clone}{\code{Classifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classifier-new"></a>}}
\if{latex}{\out{\hypertarget{method-Classifier-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classifier$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classifier-entrainer"></a>}}
\if{latex}{\out{\hypertarget{method-Classifier-entrainer}{}}}
\subsection{Method \code{entrainer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classifier$entrainer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classifier-predire"></a>}}
\if{latex}{\out{\hypertarget{method-Classifier-predire}{}}}
\subsection{Method \code{predire()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classifier$predire()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Classifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
